pyproject.toml
[project]
name = "unlockxiaomi-colombia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "django>=6.0",
    "django-allauth[socialaccount]>=65.13.1",
    "django-tailwind-cli>=4.5.1",
    "pillow>=12.1.0",
    "psycopg[binary]>=3.3.2",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
